+++ src/utils/parser/utils.js
@@ -5,8 +5,10 @@ import * as t from "babel-types";
 import toPairs from "lodash/toPairs";
 import isEmpty from "lodash/isEmpty";
 import uniq from "lodash/uniq";
+
 import { getAst, traverseAst } from "./utils/ast";
-import { isFunction } from "./utils/helpers";
+import { isFunction, isVariable } from "./utils/helpers";
+import { getClosestExpression, getClosestScope } from "./utils/closest";
 
 import type { SourceText, Location, Frame, TokenResolution } from "../../types";
 type Scope = {
@@ -17,30 +19,6 @@ type Scope = {
   parent: Scope,
   bindings: Object[]
 };
-
-function getNodeValue(node) {
-  if (t.isThisExpression(node)) {
-    return "this";
-  }
-
-  return node.name;
-}
-
-function getMemberExpression(root) {
-  function _getMemberExpression(node, expr) {
-    if (t.isMemberExpression(node)) {
-      expr = [node.property.name].concat(expr);
-      return _getMemberExpression(node.object, expr);
-    }
-
-    if (t.isThisExpression(node)) {
-      return ["this"].concat(expr);
-    }
-    return [node.name].concat(expr);
-  }
-
-  return _getMemberExpression(root, []);
-}
 
 function getScopeVariables(scope: Scope) {
   const { bindings } = scope;
@@ -59,79 +37,6 @@ function getScopeChain(scope: Scope): Scope[] {
   } while ((scope = scope.parent));
 
   return scopes;
-}
-
-/**
- * helps find member expressions on one line and function scopes that are
- * often many lines
- */
-function nodeContainsLocation({ node, location }) {
-  const { start, end } = node.loc;
-  const { line, column } = location;
-
-  if (start.line === end.line) {
-    return (
-      start.line === line && start.column <= column && end.column >= column
-    );
-  }
-
-  // node is likely a function parameter
-  if (start.line === line) {
-    return start.column <= column;
-  }
-
-  // node is on the same line as the closing curly
-  if (end.line === line) {
-    return end.column >= column;
-  }
-
-  // node is either inside the block body or outside of it
-  return start.line < line && end.line > line;
-}
-
-function isLexicalScope(path) {
-  return t.isBlockStatement(path) || isFunction(path) || t.isProgram(path);
-}
-
-function getClosestMemberExpression(source, token, location) {
-  let expression = null;
-  traverseAst(source, {
-    enter(path) {
-      const { node } = path;
-      if (
-        t.isMemberExpression(node) &&
-        node.property.name === token &&
-        nodeContainsLocation({ node, location })
-      ) {
-        const memberExpression = getMemberExpression(node);
-        expression = {
-          value: memberExpression.join("."),
-          location: node.loc
-        };
-      }
-    }
-  });
-
-  return expression;
-}
-
-export function getClosestExpression(
-  source: SourceText,
-  token: string,
-  location: Location
-) {
-  const memberExpression = getClosestMemberExpression(source, token, location);
-  if (memberExpression) {
-    return memberExpression;
-  }
-
-  const path = getClosestPath(source, location);
-  if (!path || !path.node) {
-    return;
-  }
-
-  const { node } = path;
-  return { value: getNodeValue(node), location: node.loc };
 }
 
 // Resolves a token (at location) in the source to determine if it is in scope
@@ -155,51 +60,6 @@ export function resolveToken(
     expression,
     inScope
   };
-}
-
-export function getClosestScope(source: SourceText, location: Location) {
-  const ast = getAst(source);
-  if (isEmpty(ast)) {
-    return null;
-  }
-
-  let closestPath = null;
-
-  traverse(ast, {
-    enter(path) {
-      if (
-        isLexicalScope(path) &&
-        nodeContainsLocation({ node: path.node, location })
-      ) {
-        closestPath = path;
-      }
-    }
-  });
-
-  if (!closestPath) {
-    return;
-  }
-
-  return closestPath.scope;
-}
-
-export function getClosestPath(source: SourceText, location: Location) {
-  const ast = getAst(source);
-  if (isEmpty(ast)) {
-    return null;
-  }
-
-  let closestPath = null;
-
-  traverse(ast, {
-    enter(path) {
-      if (nodeContainsLocation({ node: path.node, location })) {
-        closestPath = path;
-      }
-    }
-  });
-
-  return closestPath;
 }
 
 export function getVariablesInLocalScope(scope: Scope) {
